[EVE-7] Refactor GL.controller for clean support of different renderers.
